Fix project score calculation
This fixes a small accumulation of errors that can occur in the
last loop of the function. Since it divides by weighted_score
multiple times, very small rounding errors can accumulate and
will prevent some project types, for example, from being able to
get a score of 100 (instead, it will show 99). We can move the
division to the end, and do it only once, as this is mathematically
equivalent and doesn't accumulate errors.