Bump precision in score calculation variable
This fixes an edge case where you have a single fact type for a
project with weight 100, and a fact value with score 100 (such as
boolean True). This maximum value will exceed the NUMERIC(9,5)
size of the weighted_score variable. So we just bump the precision
here from 9 to 10. This doesn't cause downstream precision problems
because weighted_score is only used on the following line which
truncates to (9,2) on each iteration.